Split files for codeclimate
